Fixing issue with deleting control states
Since moving 'validators' to the $scope.valid object, buildValidators had been
called each time validate() was called. This is okay in most cases since the
array remains local to each scope for each call. When something is deleted, the
validators from whatever previous control remain and the validator was
attempting to validate the scope of the parent of the children that just got
deleted.

deleteChildStates now clears the validators collection from this parent scope
which prevents an erroneous invalid state from polluting the global validation
hash.